Fix circular import between prompt.py and bash.py
