+++ openhands/agenthub/codeact_agent/tools/bash.py
@@ -1,9 +1,7 @@
-import re
-import sys
-
 from litellm import ChatCompletionToolParam, ChatCompletionToolParamFunctionChunk
 
 from openhands.llm.tool_names import EXECUTE_BASH_TOOL_NAME
+from openhands.utils.prompt import refine_prompt
 
 _DETAILED_BASH_DESCRIPTION = """Execute a bash command in the terminal within a persistent shell session.
 
@@ -34,21 +32,6 @@ _SHORT_BASH_DESCRIPTION = """Execute a bash command in the terminal.
 * Long running commands: For commands that may run indefinitely, it should be run in the background and the output should be redirected to a file, e.g. command = `python3 app.py > server.log 2>&1 &`. For commands that need to run for a specific duration, you can set the "timeout" argument to specify a hard timeout in seconds.
 * Interact with running process: If a bash command returns exit code `-1`, this means the process is not yet finished. By setting `is_input` to `true`, the assistant can interact with the running process and send empty `command` to retrieve any additional logs, or send additional text (set `command` to the text) to STDIN of the running process, or send command like `C-c` (Ctrl+C), `C-d` (Ctrl+D), `C-z` (Ctrl+Z) to interrupt the process.
 * One command at a time: You can only execute one bash command at a time. If you need to run multiple commands sequentially, you can use `&&` or `;` to chain them together."""
-
-
-def refine_prompt(prompt: str):
-    if sys.platform == 'win32':
-        # Replace 'bash' with 'powershell' including tool names like 'execute_bash'
-        # First replace 'execute_bash' with 'execute_powershell' to handle tool names
-        result = re.sub(
-            r'\bexecute_bash\b', 'execute_powershell', prompt, flags=re.IGNORECASE
-        )
-        # Then replace standalone 'bash' with 'powershell'
-        result = re.sub(
-            r'(?<!execute_)(?<!_)\bbash\b', 'powershell', result, flags=re.IGNORECASE
-        )
-        return result
-    return prompt
 
 
 def create_cmd_run_tool(
